+++ src/components/vanilla/charts/PieChart/index.tsx
@@ -18,9 +18,9 @@ import { Pie, getElementAtEvent } from 'react-chartjs-2';
 
 import formatValue from '../../../util/format';
 import Container from '../../Container';
-import { useTheme } from '@embeddable.com/react';
 import { setChartJSDefaults } from '../../../util/chartjs/common';
 import { Theme } from '../../../../themes/theme';
+import { useTheme } from '@embeddable.com/react';
 
 ChartJS.register(
   ChartDataLabels,
@@ -130,7 +130,7 @@ export default (props: Props) => {
     <Container {...props} className="overflow-y-hidden">
       <Pie
         height="100%"
-        options={chartOptions(updatedProps)}
+        options={chartOptions(updatedProps, theme)}
         data={chartData(updatedProps, chartColors)}
         ref={chartRef}
         onClick={handleClick}
@@ -139,7 +139,7 @@ export default (props: Props) => {
   );
 };
 
-function chartOptions(props: Props): ChartOptions<'pie'> {
+function chartOptions(props: Props, theme: Theme): ChartOptions<'pie'> {
   return {
     responsive: true,
     maintainAspectRatio: false,
@@ -151,11 +151,13 @@ function chartOptions(props: Props): ChartOptions<'pie'> {
     plugins: {
       datalabels: {
         // Great resource: https://quickchart.io/documentation/chart-js/custom-pie-doughnut-chart-labels/
+        backgroundColor: theme.charts.pie.labels.backgroundColor,
+        borderRadius: theme.charts.pie.labels.borderRadius,
+        color: theme.charts.pie.labels.color,
         display: props.showLabels ? 'auto' : false,
-        backgroundColor: '#fff',
-        borderRadius: 8,
         font: {
-          weight: 'normal',
+          size: theme.charts.pie.labels.font.size,
+          weight: theme.charts.pie.labels.font.weight,
         },
         formatter: (v) => {
           const val = v